+++ Data/Scripts/RealEnergy/Session.cs
@@ -61,7 +61,6 @@ namespace TSUT.HeatManagement
             Instance = this;
             MyLog.Default.WriteLine($"[HeatManagement] HeatSession instance created.");
 
-            _heatApi.Registry.RegisterHeatBehaviorFactory(new VentHeatManagerFactory());
             _heatApi.Registry.RegisterHeatBehaviorFactory(new ExhaustHeatManagerFactory());
             _heatApi.Registry.RegisterHeatBehaviorFactory(new HeatPipeManagerFactory());
             _heatApi.Registry.RegisterHeatBehaviorFactory(new HeatVentManagerFactory());
